+++ app/src/main/java/com/example/cryptolize/domain/models/detailModel/MarketData.kt
@@ -57,5 +57,5 @@ data class Ath(
 )
 
 data class Sparkline7d(
-    val price: List<Double>
+    val price: List<Float>
 )
